chat/requirements.txt
flask==2.3.2
google-generativeai==0.3.2